Implement header changes in LC. Centralize the structure by depending on avail-subxt structures (still not using subxt).

+++ src/proof.rs
@@ -9,7 +9,7 @@ use tracing::{error, trace};
 // Just a wrapper function, to be used when spawning threads for verifying proofs
 // for a certain block
 fn kc_verify_proof_wrapper(
-	block_num: u64,
+	block_num: u32,
 	row: u16,
 	col: u16,
 	total_rows: usize,
@@ -32,7 +32,7 @@ fn kc_verify_proof_wrapper(
 
 /// Verifies proofs for given block, cells and commitments
 pub fn verify_proof(
-	block_num: u64,
+	block_num: u32,
 	total_rows: u16,
 	total_cols: u16,
 	cells: &[Cell],
